v1/requirements.txt
pyyaml
requests
flask
gevent
pydantic
pytest